requirements.txt
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.12.0
bleach==5.0.1
bleach-whitelist==0.0.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==38.0.4
cvelib==1.2.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-allauth==0.51.0
django-allauth-2fa @ git+https://github.com/esarneso/django-allauth-2fa.git@6bd3101916f2fca6f2ff414b3b1181bbd44d1b99
django-cors-headers==3.13.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-formtools==2.4
django-oauth-toolkit==2.2.0
django-otp==1.1.4
django-qr-code==3.1.1
django-webpack-loader==1.8.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
segno==1.5.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
wrapt==1.14.1
